+++ src/main/java/com/runninghi/runninghibackv2/post/application/service/PostKeywordService.java
@@ -1,13 +1,13 @@
-package com.runninghi.runninghibackv2.post.postkeyword.application.service;
+package com.runninghi.runninghibackv2.post.application.service;
 
 import com.runninghi.runninghibackv2.keyword.application.dto.request.KeywordRequest;
 import com.runninghi.runninghibackv2.keyword.application.dto.response.KeywordResponse;
 import com.runninghi.runninghibackv2.keyword.application.service.KeywordService;
 import com.runninghi.runninghibackv2.keyword.domain.aggregate.entity.Keyword;
 import com.runninghi.runninghibackv2.keyword.domain.service.KeywordChecker;
-import com.runninghi.runninghibackv2.post.postkeyword.domain.aggregate.entity.PostKeyword;
-import com.runninghi.runninghibackv2.post.postkeyword.domain.aggregate.vo.PostKeywordId;
-import com.runninghi.runninghibackv2.post.postkeyword.domain.repository.PostKeywordRepository;
+import com.runninghi.runninghibackv2.post.domain.aggregate.entity.PostKeyword;
+import com.runninghi.runninghibackv2.post.domain.aggregate.vo.PostKeywordId;
+import com.runninghi.runninghibackv2.post.domain.repository.PostKeywordRepository;
 import com.runninghi.runninghibackv2.post.domain.aggregate.entity.Post;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
