+++ src/main/java/com/runninghi/runninghibackv2/post/domain/repository/PostKeywordRepository.java
@@ -1,11 +1,9 @@
-package com.runninghi.runninghibackv2.post.postkeyword.domain.repository;
+package com.runninghi.runninghibackv2.post.domain.repository;
 
-import com.runninghi.runninghibackv2.bookmark.domain.aggregate.entity.Bookmark;
-import com.runninghi.runninghibackv2.post.postkeyword.domain.aggregate.entity.PostKeyword;
-import com.runninghi.runninghibackv2.post.postkeyword.domain.aggregate.vo.PostKeywordId;
+import com.runninghi.runninghibackv2.post.domain.aggregate.entity.PostKeyword;
+import com.runninghi.runninghibackv2.post.domain.aggregate.vo.PostKeywordId;
 import com.runninghi.runninghibackv2.post.domain.aggregate.entity.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
