--- src/main/java/com/runninghi/runninghibackv2/post/postkeyword/application/dto/test.java
@@ -1,4 +1,0 @@
-package com.runninghi.runninghibackv2.post.postkeyword.application.dto;
-
-public class test {
-}
