--- src/main/java/com/runninghi/runninghibackv2/post/postkeyword/infrastructure/repository/test.java
@@ -1,4 +1,0 @@
-package com.runninghi.runninghibackv2.post.postkeyword.infrastructure.repository;
-
-public class test {
-}
