--- src/main/java/com/runninghi/runninghibackv2/post/postkeyword/infrastructure/service/test.java
@@ -1,4 +1,0 @@
-package com.runninghi.runninghibackv2.post.postkeyword.infrastructure.service;
-
-public class test {
-}
